refactor OpenGL folder
