+++ Engine/Renderer/OpenGL/DeferredRenderer.h
@@ -7,7 +7,7 @@
 #include "ZeloPrerequisites.h"
 #include "ZeloGLPrerequisites.h"
 #include "ForwardRenderer.h"
-#include "GLTexture.h"
+#include "Renderer/OpenGL/Resource/GLTexture.h"
 #include <Renderer/OpenGL/ingredients/plane.h>
 #include <Renderer/OpenGL/ingredients/torus.h>
 #include <Renderer/OpenGL/ingredients/teapot.h>
