+++ Engine/Renderer/OpenGL/Resource/MeshLoader.h
@@ -7,9 +7,9 @@
 
 #include "ZeloPrerequisites.h"
 #include "ZeloGLPrerequisites.h"
-#include "GLTexture.h"
-#include "GLMesh.h"
-#include "Material.h"
+#include "Renderer/OpenGL/Resource/GLTexture.h"
+#include "Renderer/OpenGL/Resource/GLMesh.h"
+#include "Renderer/OpenGL/Resource/Material.h"
 #include "Core/ECS/Entity.h"
 
 #include <assimp/scene.h>
